Added the singleton database link into it, updated the getCity methods.
