+++ src/main/java/com/napier/sem/DatabaseLink.java
@@ -57,6 +57,7 @@ public class DatabaseLink
             {
                 // Close connection
                 con.close();
+                System.out.println("Connect Successfully closed");
             }
             catch (Exception e)
             {
@@ -66,43 +67,7 @@ public class DatabaseLink
     }
 
 
-    public City getCity(int id)
-    {
-        try
-        {
-            // Create an SQL statement
-            Statement stmt = con.createStatement();
-            // Create string for SQL statement
-            String strSelect =
-                    "SELECT id, name, countryCode, district, population "
-                            + "FROM city "
-                            + "WHERE id = " + id;
-            // Execute SQL statement
-            ResultSet rset = stmt.executeQuery(strSelect);
-            // Return new employee if valid.
-            // Check one is returned
-            if (rset.next())
-            {
-                City city = new City();
-                city.id = rset.getInt("id");
-                city.name = rset.getString("name");
-                city.countryCode = rset.getString("countryCode");
-                city.district = rset.getString("district");
-                city.population = rset.getInt("population");
 
-
-                return city;
-            }
-            else
-                return null;
-        }
-        catch (Exception e)
-        {
-            System.out.println(e.getMessage());
-            System.out.println("Failed to get city details");
-            return null;
-        }
-    }
